fix: address PR review feedback across splash screen and footer
