+++ src/utils.ts
@@ -1,24 +1,20 @@
+import { execSync } from "node:child_process"
 import { readFileSync } from "node:fs"
 import { homedir } from "node:os"
 import { dirname, resolve, sep } from "node:path"
 import { fileURLToPath } from "node:url"
 
-export function getGitBranch(): string {
+export function getGitBranch(cwd?: string): string | undefined {
 	try {
-		let gitPath = resolve(process.cwd(), ".git")
-		let head: string
-		try {
-			head = readFileSync(resolve(gitPath, "HEAD"), "utf-8").trim()
-		} catch {
-			const gitFile = readFileSync(gitPath, "utf-8").trim()
-			if (!gitFile.startsWith("gitdir: ")) return ""
-			gitPath = resolve(process.cwd(), gitFile.slice("gitdir: ".length))
-			head = readFileSync(resolve(gitPath, "HEAD"), "utf-8").trim()
-		}
-		if (head.startsWith("ref: refs/heads/")) return head.slice("ref: refs/heads/".length)
-		return head.slice(0, 7)
+		return (
+			execSync("git symbolic-ref --short HEAD", {
+				cwd: cwd ?? process.cwd(),
+				encoding: "utf8",
+				stdio: ["ignore", "pipe", "ignore"],
+			}).trim() || undefined
+		)
 	} catch {
-		return ""
+		return undefined
 	}
 }
